Created a minimized getting started page and removed old content
The getting started guide was full with stack or tool specific
information. Reading has been a pain because of the numerous links.

This change intorduces a tiny minimized getting started page with the
main information and links.
All old and unused/unlinked content was removed.
Links that were not working on the starting page have been removed.